more fixes to the way scripts are enquque'd
So I was wrong about that, the javascript wasn’t getting loaded by the
admin area it was just loaded on the login screen itself.  Since you
aren’t logged in yet…  So the fix to prevent js from being applied to
the login page is to detect if you’re on the login page and put that
into the condition with the is_admin bit…  Also moved the add_action
part out of the if statement again since it was actually in the right
spot to begin with.